refactor and feat: sprint 11

+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Genre {
-    private long id;
+
+    private int id;
     private String name;
 }
